feat: Add slash commands (#105)
# New Features
- Added new class Respondable. What it does? It is for sending replies with same method for Message and Interaction classes. https://github.com/hiro-team/hiro-bot/pull/105/commits/86f86dd4547ed4ad2a2592f6262141d1c105bd35
- Some commands are updated for slash commands. Added $options property to them. https://github.com/hiro-team/hiro-bot/pull/105/commits/c40c890717727e9c75389545b44bad331fa6e30e
- CommandLoader class updates for automatic saving slash commands https://github.com/hiro-team/hiro-bot/pull/105/commits/cafcf88c1a5c9c34d3c38bb6b3cb3a46c515d7fd
- Changed commands' $msg and $args variable types https://github.com/hiro-team/hiro-bot/pull/105/commits/2a25a044c6ecd6cff30b7ac031b06f8bf2da5107
- Added latest changes from master #104
- Added slash commands registration system. There is a variable named 'allowed_categories'. If command's category is allowed. It will be registered. If not it will be deleted from slash commands. https://github.com/hiro-team/hiro-bot/pull/105/commits/d136a2033c45449f6915d92fee442d5ddb855dcf
## Removed Features
- RPG commands has been deleted. https://github.com/hiro-team/hiro-bot/pull/105/commits/d201d9ec659c11a46be46fb99c637056cd56281f

+++ src/commands/utility/Clear.php
@@ -20,7 +20,9 @@
 
 namespace hiro\commands;
 
+use Discord\Helpers\Collection;
 use Discord\Parts\Embed\Embed;
+use Discord\Parts\Interactions\Command\Option;
 
 /**
  * Clear
@@ -38,6 +40,13 @@ class Clear extends Command
         $this->description = "Clears messages";
         $this->aliases = ["purge"];
         $this->category = "utility";
+        $this->options = [
+            (new Option($this->discord))
+                ->setType(Option::INTEGER)
+                ->setName('amount')
+                ->setDescription('Amount of messages to clear')
+                ->setRequired(true)
+        ];
     }
 
     /**
@@ -56,17 +65,22 @@ class Clear extends Command
             $embed->setDescription($language->getTranslator()->trans('commands.clear.no_perm'));
             $embed->setColor("#ff000");
             $embed->setTimestamp();
-            $msg->channel->sendEmbed($embed);
+            $msg->reply($embed);
             return;
         }
-        $limit = $args[0];
+        if($args instanceof Collection && $args->get('name', 'amount') !== null) {
+            $limit = $args->get('name', 'amount')->value;
+        } else if (is_array($args)) {
+            $limit = $args[0] ?? null;
+        }
+        $limit ??= null;
         if (!isset($limit)) {
             $embed = new Embed($this->discord);
             $embed->setTitle($language->getTranslator()->trans('commands.clear.error'));
             $embed->setDescription($language->getTranslator()->trans('commands.clear.no_amount'));
             $embed->setColor("#ff000");
             $embed->setTimestamp();
-            $msg->channel->sendEmbed($embed);
+            $msg->reply($embed);
             return;
         } else if (!is_numeric($limit)) {
             $embed = new Embed($this->discord);
@@ -74,7 +88,7 @@ class Clear extends Command
             $embed->setDescription($language->getTranslator()->trans('commands.clear.no_numeric_arg'));
             $embed->setColor("#ff000");
             $embed->setTimestamp();
-            $msg->channel->sendEmbed($embed);
+            $msg->reply($embed);
             return;
         } else if ($limit < 1 || $limit > 100) {
             $embed = new Embed($this->discord);
@@ -82,18 +96,20 @@ class Clear extends Command
             $embed->setDescription($language->getTranslator()->trans('commands.clear.limit'));
             $embed->setColor("#ff000");
             $embed->setTimestamp();
-            $msg->channel->sendEmbed($embed);
+            $msg->reply($embed);
             return;
         }
-        $msg->channel->limitDelete($limit)->then(function() use ($msg, $limit, $language) {
+        $msg->channel->limitDelete($limit)->then(function () use ($msg, $limit, $language) {
             $embed = new Embed($this->discord);
             $embed->setTitle("Clear Command");
             $embed->setDescription(sprintf($language->getTranslator()->trans('commands.clear.deleted'), $limit));
             $embed->setColor("#5558E0");
             $embed->setTimestamp();
-            $msg->channel->sendEmbed($embed)->then(function ($msg) {
+            $msg->reply($embed)->then(function ($msg) {
                 $this->discord->getLoop()->addTimer(3.0, function () use ($msg) {
-                    $msg->delete();
+                    if($msg) {
+                        $msg->delete();
+                    }
                 });
             });
         }, function (\Throwable $reason) use ($msg, $language) {
